BUG: Recursive event handling crash caused by mutually triggering controls. Solution: Use a variable to check if an event was caused by user intervention.
Details: The yaw and pitch sliders are linked to the section plane. When the sliders are adjusted
the section plane is updated; when the section plane is dragged, the sliders are updated. Since
we set the values of the sliders using Mootols .set() method, this automatically triggers events
onChange() and onComplete(), whether it was triggered by the user, or by adjustments on the section plane.
By using the above variable, we can ignore these events if it was not triggered by the user.


git-svn-id: svn://svn.code.sf.net/p/ngembryo/code@29 b13ead23-0dcf-405c-9d34-bbe0b799025b